server/requirements.txt
cloud-sql-python-connector==1.2.2
Flask==2.2.3
Flask-Cors==3.0.10
GDAL==3.6.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
pycparser==2.21
PyMySQL==1.0.3
SQLAlchemy==2.0.12